1. Allowing a filter template to be applied without needing to patch.
2. Updated Renard module.  It's not the only way to implement it, especially considering the double-dispatching, but it is a way.

3. Some ancillary unit changes to get a few more modules to compile; due to core framework refactor.

+++ Modules/Input/DirectXJoystick/Joystick.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
+using DirectXJoystick;
 using SlimDX.DirectInput;
 using Vixen.Module.Input;
 using VixenModules.Input.DirectXJoystick.Input;
@@ -94,7 +95,7 @@ namespace VixenModules.Input.DirectXJoystick {
 				device = new SlimDX.DirectInput.Joystick(directInput, deviceInstance.InstanceGuid);
 				// This needs System.Windows.Forms.Control.
 				device.SetCooperativeLevel(IntPtr.Zero, CooperativeLevel.Nonexclusive | CooperativeLevel.Background);
-				device.Properties.SetRange(Position.MinValue, Position.MaxValue);
+				device.Properties.SetRange((int)Position.MinValue, (int)Position.MaxValue);
 
 				Inputs = _GetInputs(device).ToArray();
 			}
